v0.4.0:#4: Adds two random number range engines based on xorshift.
Adds a new xorshift range engine implementation with 32 and 64 bit versions.
The range engine generates random numbers by filling a provided span of bytes
or integers, using an xorshift algorithm with multiple seeds and rotation
to improve randomness.

Also, the uniform random number generator and xorshift engine headers have been
moved to a new `rng` subdirectory, and a dedicated test suite is added for
the range engines in a `rnrg` subdirectory.